* release:
  Minor changes to resolve Visual Studio warnings

+++ maptk/transform.cxx
@@ -218,7 +218,7 @@ necker_reverse(vital::camera_map_sptr& cameras,
   {
     lc += p.second->loc();
   }
-  lc /= lms.size();
+  lc /= static_cast<double>(lms.size());
 
   // compute the mean camera center
   vital::vector_3d cc(0.0, 0.0, 0.0);
@@ -226,7 +226,7 @@ necker_reverse(vital::camera_map_sptr& cameras,
   {
     cc += p.second->center();
   }
-  cc /= cams.size();
+  cc /= static_cast<double>(cams.size());
 
   vital::vector_3d axis(cc - lc);
   axis.normalize();
